+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   * Renamed MPGeneralContorller.thConfigFilePath to mpConfigFilePath.
   * Renamed MPSettingsController.getTYPE -> getTYPEWithDefault.
   * Created MPSettingsController.getTYPEIfSet() that returns null if not set.
+  * Created MPBaseTherionRunner class with getUserDefinedTherionExecutablePath() method.
 
 ## 0.2.36 - 2026-02-24 - The [Danaids](https://en.wikipedia.org/wiki/Danaids) release
 * New features: